Refactor code into ES modules for better maintainability and modularity
